Updated response and added api for schedule

+++ src/main/java/com/harman/ebook/vaccination/covid/constants/LovConstants.java
@@ -12,8 +12,8 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 public final class LovConstants {
-  public static final int LOV_APP_STATUS_BOOKED = 1;
-  public static final int LOV_APP_STATUS_CANCELED = 2;
-  public static final int LOV_APP_STATUS_RESCHEDULED = 3;
-  public static final int LOV_APP_STATUS_COMPLETED = 4;
+  public static final Short LOV_APP_STATUS_BOOKED = 1;
+  public static final Short LOV_APP_STATUS_CANCELED = 2;
+  public static final Short LOV_APP_STATUS_RESCHEDULED = 3;
+  public static final Short LOV_APP_STATUS_COMPLETED = 4;
 }
